+++ packages/blade/src/components/Drawer/__tests__/Drawer.test.stories.tsx
@@ -210,6 +210,5 @@ export default {
     },
     a11y: { disable: true },
     essentials: { disable: true },
-    actions: { disable: true },
   },
 };