+++ packages/blade/src/components/Tooltip/__tests__/Tooltip.test.stories.tsx
@@ -185,6 +185,5 @@ export default {
     },
     a11y: { disable: true },
     essentials: { disable: true },
-    actions: { disable: true },
   },
 };