+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @razorpay/blade
 
+## 12.25.0
+
+### Minor Changes
+
+- 59fb0163: feat(PhoneNumberInput): add virtualization to phone number input
+
+## 12.24.0
+
+### Minor Changes
+
+- f7b868ad: feat(blade): add full page modal
+
+### Patch Changes
+
+- 9eb42df8: fix(StepGroup): extra space because of trailing property
+
 ## 12.23.1
 
 ### Patch Changes